もともと登録していたコンテンツを料金をグレードアップした際に消えないように修正

+++ app/models/user.rb
@@ -45,10 +45,10 @@ class User < ActiveRecord::Base
   has_many :user_plans
 
   def self.delete_dependent(current_user)
+    Subscription.where(user_id: current_user.id).delete_all if current_user.payments.last.amount == 2000
     CreditCard.where(user_id: current_user.id).delete_all
     UserPlan.where(user_id: current_user.id).delete_all
     Payment.where(user_id: current_user.id).delete_all
-    Subscription.where(user_id: current_user.id).delete_all
   end
 
   def self.find_for_google_oauth2(access_token, _signed_in_resourse = nil)
